ADD : 광고부분에 동영상 추가
